Rewrite raw candidate/display string handling.
Should help with issue #10.

Now, there is a hash table, `helm-company-display-candidates-hash`, whose keys
are display strings (with NO properties), and whose values
are (FORMATTED-DISPLAY . COMPANY-CANDIDATE) pairs. COMPANY-CANDIDATE is the
original completion candidate as supplied by Company, properties and all;
FORMATTED-DISPLAY is the display string (with font-lock properties, etc) for
Helm to display.